Update macos actions (#7845)
* ci: move macos-13 actions to macos-14

We have some GitHub Actions which we try to run on the oldest version of
macOS available. That's currently macos-13, but GitHub is deprecating it:

> This is a scheduled macos-13 brownout.
> The macOS-13 based runner images > are being deprecated. For more details,
> see https://github.com/actions/runner-images/issues/13046.

We last bumped this from macos-12 to macos-13 in 8b68332 #7275

* ci: do not specify architectures in macOS CI builds

GitHub has is phasing out su pport for macOS x86_64 runners

* ci: fold the macos-X-from-tarball jobs into a single job

* ci: fold macos-26 and macos-14 jobs into a single job

* Use macos-15-intel runners instead of macos-14 (arm64).

If we test compilation two times, at least do it on different host
archs.
Also macOS 15 supports latest Xcode with macOS 26.1 SDK.

* Make job names more meaningful

* Fix indentation in macos-xcodebuild-universal

---------

Co-authored-by: Charles Kerr <charles@charleskerr.com>